backend/requirement.txt
fastapi
uvicorn
python-multipart
python-dotenv
requests
langgraph
langchain
pydantic
google-api-python-client
python-pptx